Refactor: Deprecate `AppWidgetScreen.kt` in favor of `WidgetScreen.kt`
This commit refactors the widget picker functionality by deprecating and removing the old `AppWidgetScreen.kt` file. Its logic, previously responsible for handling widget selection and drag initiation, has been fully consolidated into the existing `WidgetScreen.kt`.

This change streamlines the codebase by removing redundant composables and centralizing all widget-picking UI and state management into a single, more robust file.

### Core Changes:

*   **File Deletion:** The file `AppWidgetScreen.kt` has been deleted.
*   **Consolidation:** The composable `EblanAppWidgetProviderInfoItem`, which handled rendering and drag-and-drop logic for individual widgets in the picker, has been removed from the old file. This functionality is now entirely managed within `WidgetScreen.kt`.
*   **Package & Import Updates:**
    *   The package name for the widget screen has been unified from `...screen.appwidget` to `...screen.widget`.
    *   Imports across `HorizontalPagerScreen.kt` and `ApplicationScreen.kt` have been updated to point to the consolidated `AppWidgetScreen` composable within the new `widget` package.
*   **UI Text Cleanup:** A minor text label in `WidgetScreen.kt` was simplified to be more user-friendly.

### Drag-and-Drop Improvements:

*   **Increased Edge Detection Area:** The drag detection area for triggering a page turn in both the main grid (`DragGridItemHelper.kt`) and folder view (`DragGridItemHelper.kt`) has been increased from `15.dp` to `30.dp`. This makes it easier for users to scroll between pages while dragging an item near the screen's edge.
