+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/drag/DragGridItemHelper.kt
@@ -111,7 +111,7 @@ internal suspend fun handleDragGridItem(
     }
 
     val edgeDistance = with(density) {
-        15.dp.roundToPx()
+        30.dp.roundToPx()
     }
 
     val horizontalPadding = leftPadding + rightPadding
