+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/folderdrag/DragGridItemHelper.kt
@@ -79,7 +79,7 @@ internal suspend fun handleDragFolderGridItem(
     }
 
     val edgeDistance = with(density) {
-        15.dp.roundToPx()
+        30.dp.roundToPx()
     }
 
     val horizontalPadding = leftPadding + rightPadding
